+++ src/components/Countdown.tsx
@@ -2,7 +2,7 @@
 
 import React, { useEffect, useState } from 'react';
 
-import Icon from './Icon';
+import Icon from '@/lib/components/Icon';
 
 const ZEROS = {
   days: 0,
