+++ src/lib/components/Icon.tsx
@@ -1,4 +1,4 @@
-import { classes } from '@/utils/classes';
+import { classes } from '@/lib/utils/classes';
 
 export default function Icon({ name, className }: { name: string; className?: string }) {
   return <i className={classes('material-symbols-rounded', className)}>{name}</i>;
