+++ src/lib/graphql/gloabls.ts
