+++ src/lib/graphql/pages.ts
